rename dirs.
