+++ implement/io/input.go
